requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pymupdf==1.24.10
pandas==2.2.3
openpyxl==3.1.5
anthropic==0.39.0
pydantic==2.9.2
jinja2==3.1.4
python-dotenv==1.0.1
pillow==10.4.0
shapely==2.0.6
numpy==1.26.4
rapidfuzz==3.10.0
pytest==8.3.3
